EPBDS-8677 Make boarders of IntRange wider (Long)

+++ DEV/org.openl.rules/src/org/openl/rules/dt/type/CharRangeAdaptor.java
@@ -18,7 +18,7 @@ public final class CharRangeAdaptor implements IRangeAdaptor<CharRange, Characte
             return null;
         }
 
-        int max = range.getMax();
+        long max = range.getMax();
 
         if (max != Character.MAX_VALUE) {
             max = max + 1;
